+++ book.rb
@@ -4,7 +4,7 @@ class Book
   attr_reader :rentals
   attr_accessor :title, :author
 
-  def initialize(title, author)
+  def initialize(title:, author:)
     @title = title
     @author = author
     @rentals = []
