+++ person.rb
@@ -5,7 +5,7 @@ class Person
   attr_reader :id, :rentals
   attr_accessor :name, :age
 
-  def initialize(age, name = 'Unknown', parent_permission: true)
+  def initialize(age:, name: 'Unknown', parent_permission: true)
     @id = Time.now.to_i + rand(100)
     @name = name
     @age = age
