+++ printer.rb
@@ -1,0 +1,3 @@
+def l_printer(str)
+  print str
+end
