+++ student.rb
@@ -1,7 +1,7 @@
 require_relative './person'
 
 class Student < Person
-  def initialize(classroom:, age:, name: 'Unknown', parent_permission: true)
+  def initialize(age:, classroom:, name: 'Unknown', parent_permission: true)
     super(age: age, name: name, parent_permission: parent_permission)
     @classroom = classroom
   end
